consolidate static content pages; address PR #93 comments
