+++ client/src/pages/about_wepp/About_WEPP.css
@@ -66,3 +66,69 @@
   padding: 3px 6px;
   border-radius: 4px;
 }
+
+hr {
+  margin-top: 30px;
+}
+
+.text-center {
+  text-align: center;
+}
+
+.dash {
+  font-size: 3em;
+  font-weight: 100;
+}
+
+#wepp-container-main {
+  width: 100%;
+  text-align: left;
+  color: white;
+  background: #333;
+  padding: 0 40px 50px 40px;
+}
+
+#wepp-container-main h2 {
+  text-align: center;
+  font-size: 2em;
+  font-weight: 800;
+  padding-top: 30px;
+  margin-bottom: 30px;
+}
+
+#wepp-container-main h3 {
+  text-align: left;
+  font-size: 1.7em;
+  font-weight: 800;
+  padding-top: 20px;
+  margin-bottom: 20px;
+}
+
+#wepp-container-main h4 {
+  font-size: 1.3em;
+  font-weight: 800;
+  padding-top: 20px;
+  margin-bottom: 20px;  
+}
+
+#wepp-container-main p, #wepp-container-main ul {
+  margin-bottom: 24px;
+}
+
+#wepp-container-main ul {
+  padding-left: 50px;
+}
+
+#wepp-container-main ul li span {
+  font-weight: 800;
+  font-style: italic;
+}
+
+#wepp-container-main section {
+  padding: 40px 40px 0 40px;
+}
+
+#wepp-container-main img {
+  width: 100%;
+  padding-top: 40px;
+}
