+++ client/src/pages/team/Team.tsx
@@ -1,10 +1,292 @@
 import React from 'react';
 import { useIsSmallScreen } from '../../hooks/useIsSmallScreen';
-import TeamSidePanelContent from '../../components/side-panels/team-side-panel/TeamSidePanel';
 import SmallScreenNotice from '../../components/small-screen-notice/SmallScreenNotice';
-//import BottomPanel from '../../components/bottom-panels/BottomPanel';
-import TeamMain from '../../components/team-main/TeamMain';
+import RogerLew from '../../assets/images/roger_lew.png'
+import MarianaDobre from '../../assets/images/mariana_dobre.png'
+import ErinBrooks from '../../assets/images/erin_brooks.png'
+import SubhankarDas from '../../assets/images/subhankar_das.jpeg'
+import ErinHanan from '../../assets/images/erin_hanan.jpg'
+import WilliamBurke from '../../assets/images/william_burke.png'
+import LawrenceAlawode from '../../assets/images/lawrence_alawode.png'
+import MingliangLiu from '../../assets/images/mingliang_liu.png'
+import JuliePadowski from '../../assets/images/julie_padowski.png'
+import JennyAdam from '../../assets/images/jenny_adam.png'
+import KevinBladon from '../../assets/images/kevin_bladon.png'
+import RyanCole from '../../assets/images/ryan_cole.png'
+import PeteRobichaud from '../../assets/images/pete_robichaud.png'
 import './Team.css';
+
+
+const scrollToUniversity = (e: React.MouseEvent, name: string) => {
+  e.preventDefault();
+  const id = name.replace(/\s+/g, '_');
+  const element = document.getElementById(id);
+  
+  if (element) {
+      let scrollContainer = element.parentElement;
+      while (scrollContainer) {
+          const hasOverflow = window.getComputedStyle(scrollContainer).overflow;
+          if (hasOverflow === 'auto' || hasOverflow === 'scroll' || 
+              scrollContainer.scrollHeight > scrollContainer.clientHeight) {
+              break;
+          }
+          scrollContainer = scrollContainer.parentElement;
+      }
+      
+      if (scrollContainer) {
+          const navbarHeight = 60;
+          const containerRect = scrollContainer.getBoundingClientRect();
+          const elementRect = element.getBoundingClientRect();
+          const relativeTop = elementRect.top - containerRect.top + scrollContainer.scrollTop;
+          const offsetPosition = relativeTop - navbarHeight;
+          
+          scrollContainer.scrollTo({
+              top: offsetPosition,
+              behavior: 'smooth'
+          });
+      }
+  }
+}
+
+/* TEAM: SIDE PANEL CONTENT */
+export function TeamSidePanelContent() {
+  return (
+    <div className="team-panel">
+      <h2>Project Team & Partners</h2>
+      <p>
+        Select from the following to see our team of experts and partner institutions:
+      </p>
+      <br />
+
+      <button
+        className='actionButton'
+        aria-label='Washington State University'
+        title='Washington State University'
+        id='wsu'
+        onClick={(e) => scrollToUniversity(e, "Washington_State_University")}
+      >
+        Washington State University
+      </button>
+
+      <button
+        className='actionButton'
+        aria-label='University of Idaho'
+        title='University of Idaho'
+        id='ui'
+        onClick={(e) => scrollToUniversity(e, "University_of_Idaho")}
+      >
+        University of Idaho</button>
+
+      <button
+        className='actionButton'
+        aria-label='University of Nevada, Reno'
+        title='University of Nevada, Reno'
+        id='unr'
+        onClick={(e) => scrollToUniversity(e, "University_of_Nevada,_Reno")}
+      >
+        University of Nevada, Reno
+      </button>
+      
+      <button
+        className='actionButton'
+        aria-label='Oregon State University'
+        title='Oregon State University'
+        id='osu'
+        onClick={(e) => scrollToUniversity(e, "Oregon_State_University")}
+      >
+        Oregon State University
+      </button>
+
+      <button
+        className='actionButton'
+        aria-label='US Forest Service'
+        title='US Forest Service'
+        id='usfs'
+        onClick={(e) => scrollToUniversity(e, "US_Forest_Service,_Rocky_Mtn_Research_Station")}
+      >
+        US Forest Service
+      </button>
+
+      <button
+        className='actionButton'
+        aria-label='Partners'
+        title='Partners'
+        id='partners'
+        onClick={(e) => scrollToUniversity(e, "partner_section")}
+      >
+        Partners
+      </button>
+      
+      <br /><br />
+
+    </div>
+  )
+}
+
+
+/* TEAM: MAIN CONTENT */
+interface TeamMember {
+  name: string;
+  img: string;
+  univ: string;
+  role: string;
+  web: string;
+}
+
+export function TeamMainContent() {
+  const teamMembers: TeamMember[] = [
+    {
+      name: "Mingliang Liu",
+      img: MingliangLiu,
+      univ: "Washington State University",
+      role: "Assistant Research Professor, Civil & Environmental Engineering",
+      web: "https://ce.wsu.edu/faculty/liu-mingliang/"
+    },
+    {
+      name: "Julie Padowski",
+      img: JuliePadowski,
+      univ: "Washington State University",
+      role: "Research Associate Professor, School of the Environment",
+      web: "https://environment.wsu.edu/faculty/wsu-profile/julie.padowski/"
+    },
+    {
+      name: "Jenny Adam",
+      img: JennyAdam,
+      univ: "Washington State University",
+      role: "Professor, Civil & Environmental Engineering",
+      web: "https://ce.wsu.edu/faculty/adam/"
+    },
+    {
+      name: "Roger Lew",
+      img: RogerLew,
+      univ: "University of Idaho",
+      role: "Research Associate Professor, Virtual Technology and Design",
+      web: "https://www.uidaho.edu/people/rogerlew"
+    },
+    {
+      name: "Mariana Dobre",
+      img: MarianaDobre,
+      univ: "University of Idaho",
+      role: "Assistant Professor, Soil and Water Systems",
+      web: "https://www.uidaho.edu/people/mdobre"
+    },
+    {
+      name: "Erin Brooks",
+      img: ErinBrooks,
+      univ: "University of Idaho",
+      role: "Professor, Soil and Water Systems",
+      web: "https://www.uidaho.edu/people/ebrooks"
+    },
+    {
+      name: "Subhankar Das",
+      img: SubhankarDas,
+      univ: "University of Idaho",
+      role: "Postdoctoral Fellow, Soil and Water Systems",
+      web: "https://scholar.google.com/citations?user=K2ZoamkAAAAJ&hl=en"
+    },
+    {
+      name: "Erin Hanan",
+      img: ErinHanan,
+      univ: "University of Nevada, Reno",
+      role: "Associate Professor, Fire & Ecosystem Ecology",
+      web: "https://www.unr.edu/nres/people/hanan-erin"
+    },
+    {
+      name: "William Burke",
+      img: WilliamBurke,
+      univ: "University of Nevada, Reno",
+      role: "Ecohydrologic Researcher",
+      web: "https://www.wdburke.com/"
+    },
+    {
+      name: "Lawrence Alawode",
+      img: LawrenceAlawode,
+      univ: "University of Nevada, Reno",
+      role: "Doctoral Student, Hydrologic Sciences",
+      web: "https://www.unr.edu/hydrologic-sciences/people/students/lawrence-gbenga-alawode"
+    },
+    {
+      name: "Kevin Bladon",
+      img: KevinBladon,
+      univ: "Oregon State University",
+      role: "Faculty, Forest Ecosystems & Society",
+      web: "https://directory.forestry.oregonstate.edu/people/bladon-kevin"
+    },
+    {
+      name: "Ryan Cole",
+      img: RyanCole,
+      univ: "Oregon State University",
+      role: "Faculty Research Assistant, Wildfire and Water Security",
+      web: "https://www.researchgate.net/profile/Ryan-Cole-9"
+    },
+    {
+      name: "Pete Robichaud",
+      img: PeteRobichaud,
+      univ: "US Forest Service, Rocky Mtn Research Station",
+      role: "Research Engineer, Erosion Modeling and Mitigation, Wildfires",
+      web: "https://scholar.google.com/citations?user=wy3ols4AAAAJ&hl=en"
+    }
+  ];
+
+  const groupedMembers = teamMembers.reduce((acc, member) => {
+    const { univ } = member;
+    if (!acc[univ]) {
+      acc[univ] = [];
+    }
+    acc[univ].push(member);
+    return acc;
+  }, {} as Record<string, TeamMember[]>);
+
+
+  return (
+    <div id="team-container-main" className="scroll-container">  
+      <div>
+        {Object.entries(groupedMembers).map(([university, members]) => (
+          <section key={university} id={university.replace(/\s+/g, '_')}>
+            <h3 className="univ-heading">{university}</h3>
+            {members.map((person, index) => (
+              <div key={index} className="member-card">
+                <img 
+                  src={person.img} 
+                  alt={person.name} 
+                  className="w-full h-48 object-cover"
+                />
+                <div className="p-4">
+                  <h4 className="text-lg font-semibold text-gray-900"><a href={person.web} target="_blank">{person.name}</a></h4>
+                  <p className="text-gray-600">{person.role}</p>
+                </div>
+              </div>
+            ))}      
+          </section>
+        ))}
+      </div>
+      
+      <div id="partner_section">
+        <h2 className="univ-heading">Partner Institutions</h2>
+        <div className="partner-col">
+          <p className="bold">Pacific Northwest water utilities:</p>
+          <ul>
+            <li><a href="https://www.seattle.gov/utilities" target="_blank">Seattle Public Utilities</a></li>
+            <li><a href="https://www.portland.gov/water" target="_blank">Portland Water Bureau</a></li>
+            <li><a href="https://www.eweb.org/" target="_blank">Eugene Water &amp; Electric Board</a></li>
+            <li><a href="https://www.cityofsalem.net/community/household/water-utilities" target="_blank">City of Salem</a></li>
+            <li><a href="https://www.bremertonwa.gov/524/Utility-Billing" target="_blank">City of Bremerton</a></li>
+            <li><a href="https://www.medfordwater.org/" target="_blank">Medford Water Commission</a></li>
+            <li><a href="https://www.clackamasproviders.org/" target="_blank">Clackamas River Water Providers</a></li>
+            <li><a href="https://www.victoria.ca/home-property/utilities/water-system" target="_blank">City of Victoria, Canada</a></li>
+          </ul>
+        </div>
+        <div className="partner-col">
+          <p><a href="https://doh.wa.gov/community-and-environment/drinking-water/office-drinking-water" target="_blank">Washington Department of Health – Office of Drinking Water</a></p>
+          <br />
+          <p><a href="https://www.waterrf.org/" target="_blank">The Water Research Foundation</a></p>
+        </div>   
+      </div>
+      <br /><br /><br />
+    </div>
+  )
+}
+
 
 /**
  * SidePanel component
@@ -18,7 +300,7 @@ function SidePanel({ children }: { children: React.ReactNode }) {
 }
 
 /**
- * The main layout for the Team page.
+ * Layout for the TEAM page.
  */
 export default function Team() { 
   const isSmallScreen = useIsSmallScreen();
@@ -33,9 +315,8 @@ export default function Team() {
         <TeamSidePanelContent />
       </SidePanel>
       <div className='team-wrapper' style={{ position: 'relative' }}>
-        <TeamMain />
+        <TeamMainContent />
       </div>
     </div>
   )
-  
 }
